chore: upload to cdn scripts

+++ utils/scripts/upload_to_cdn.sh
@@ -1,11 +1,25 @@
 #!/bin/bash
 
+# Path to lerna.json
+LERNA_JSON="lerna.json"
+
+# Check if lerna.json exists
+if [ ! -f "$LERNA_JSON" ]; then
+  echo "lerna.json not found!"
+  exit 1
+fi
+
+# Read and process lerna.json using jq
+PACKAGE_VERSION=$(jq -r '.version' $LERNA_JSON)
+
+echo "Lerna version: $PACKAGE_VERSION"
+
 # Get the published package name
-PUBLISHED_PACKAGE=$1
+PUBLISHED_PACKAGE=$1@$PACKAGE_VERSION
+echo "Uploading published package: $PUBLISHED_PACKAGE"
 #PUBLISHED_PACKAGE="${{ steps.publish.outputs.id }}"
 
 mkdir -p ./tmp \
-  && sleep 60 \
   && npm install --prefix ./tmp "$PUBLISHED_PACKAGE" \
   && cd ./tmp/node_modules
